refactor(envoy-filter): move istio_header code to common package
