+++ privacy-profile-composer/pkg/envoy_filter/common/istio_header_test.go
@@ -1,4 +1,4 @@
-package main
+package common
 
 import (
 	"testing"
